parse cell fill
