+++ ShapeCrawler/.editorconfig
@@ -29,25 +29,26 @@ dotnet_diagnostic.SA1028.severity = none
 dotnet_diagnostic.SA1117.severity = error
 dotnet_diagnostic.SA1124.severity = none
 dotnet_diagnostic.SA1200.severity = none
-dotnet_diagnostic.SA1611.severity = none
-dotnet_diagnostic.SA1618.severity = error
-dotnet_diagnostic.SA1615.severity = none
 dotnet_diagnostic.SA1413.severity = none
-dotnet_diagnostic.SA1633.severity = none
 dotnet_diagnostic.SA1201.severity = error
 dotnet_diagnostic.SA1003.severity = error
 dotnet_diagnostic.SA1005.severity = error
 dotnet_diagnostic.SA1101.severity = error
 dotnet_diagnostic.SA1202.severity = error
 dotnet_diagnostic.SA1300.severity = error
-dotnet_diagnostic.SA1401.severity = error
+dotnet_diagnostic.SA1306.severity = error
+dotnet_diagnostic.SA1401.severity = none
 dotnet_diagnostic.SA1503.severity = error
 dotnet_diagnostic.SA1507.severity = none
 dotnet_diagnostic.SA1513.severity = error
 dotnet_diagnostic.SA1516.severity = error
 dotnet_diagnostic.SA1600.severity = error
 dotnet_diagnostic.SA1602.severity = error
+dotnet_diagnostic.SA1611.severity = none
+dotnet_diagnostic.SA1615.severity = none
+dotnet_diagnostic.SA1618.severity = error
 dotnet_diagnostic.SA1623.severity = error
+dotnet_diagnostic.SA1633.severity = none
 
 # CSXXXX
 dotnet_diagnostic.CS0649.severity = error
