+++ ShapeCrawler/AutoShapes/SCFont.cs
@@ -213,7 +213,7 @@ internal class SCFont : IFont
     {
         i = -1;
         var placeholder = shape.Placeholder as Placeholder;
-        var referencedShape = placeholder?.ReferencedShape.Value as SlideAutoShape;
+        var referencedShape = placeholder?.ReferencedShape.Value as SCAutoShape;
         var fontDataPlaceholder = new FontData();
         if (referencedShape != null)
         {
