+++ ShapeCrawler/Drawing/ShapeFill/CellFill.cs
@@ -1,0 +1,11 @@
+﻿using DocumentFormat.OpenXml;
+
+namespace ShapeCrawler.Drawing.ShapeFill;
+
+internal class CellFill : ShapeFill
+{
+    internal CellFill(SlideObject slideObject, TypedOpenXmlCompositeElement cellProperties)
+        : base(slideObject, cellProperties)
+    {
+    }
+}
